Make MWSE and CSSE share BaseObject and GameFile structs
