+++ CSSE/CSActor.h
@@ -5,6 +5,25 @@
 #include "NIIteratedList.h"
 #include "NITArray.h"
 
+namespace TES3 {
+	struct ActorVirtualTable : ObjectVirtualTable {
+		void* unknown_0x138;
+		void* unknown_0x13C;
+		void* unknown_0x140;
+		void* unknown_0x144;
+		void* unknown_0x148;
+		void* unknown_0x14C;
+		void* unknown_0x150;
+		bool(__thiscall* hasMovementSwim)(const se::cs::Object*); // 0x154
+		bool(__thiscall* hasMovementWalk)(const se::cs::Object*); // 0x154
+		bool(__thiscall* hasMovementFly)(const se::cs::Object*); // 0x154
+		void* unknown_0x160;
+		void* unknown_0x164;
+		void* unknown_0x168;
+	};
+	static_assert(sizeof(ActorVirtualTable) == 0x16C, "Actor virtual table failed size validation");
+}
+
 namespace se::cs {
 	namespace ActorFlag {
 		typedef unsigned int value_type;
@@ -16,7 +35,7 @@ namespace se::cs {
 
 	struct ItemData {
 		int count; // 0x0
-		BaseObject* owner; // 0x4
+		TES3::BaseObject* owner; // 0x4
 		union {
 			long requiredRank;
 			GlobalVariable* requiredVariable;
@@ -46,23 +65,6 @@ namespace se::cs {
 	};
 	static_assert(sizeof(Inventory) == 0x1C, "Inventory failed size validation");
 
-	struct Actor_VirtualTable : Object_VirtualTable {
-		void* unknown_0x138;
-		void* unknown_0x13C;
-		void* unknown_0x140;
-		void* unknown_0x144;
-		void* unknown_0x148;
-		void* unknown_0x14C;
-		void* unknown_0x150;
-		bool(__thiscall* hasMovementSwim)(const Object*); // 0x154
-		bool(__thiscall* hasMovementWalk)(const Object*); // 0x154
-		bool(__thiscall* hasMovementFly)(const Object*); // 0x154
-		void* unknown_0x160;
-		void* unknown_0x164;
-		void* unknown_0x168;
-	};
-	static_assert(sizeof(Actor_VirtualTable) == 0x16C, "Actor virtual table failed size validation");
-
 	struct Actor : AnimatedObject {
 		unsigned int actorFlags; // 0x4C
 		int unknown_0x50;
@@ -72,15 +74,15 @@ namespace se::cs {
 		int unknown_0x78;
 
 		inline bool hasMovementSwim() const {
-			return vtbl.actor->hasMovementSwim(this);
+			return vTable.actor->hasMovementSwim(this);
 		}
 
 		inline bool hasMovementWalk() const {
-			return vtbl.actor->hasMovementWalk(this);
+			return vTable.actor->hasMovementWalk(this);
 		}
 
 		inline bool hasMovementFly() const {
-			return vtbl.actor->hasMovementFly(this);
+			return vTable.actor->hasMovementFly(this);
 		}
 	};
 	static_assert(sizeof(Actor) == 0x7C, "CS::Actor failed size validation");
