+++ MWSE/TES3Object.cpp
@@ -66,535 +66,6 @@
 #include "Log.h"
 
 namespace TES3 {
-	void* BaseObject::operator new(size_t size) {
-		return se::memory::_new(size);
-	}
-
-	void BaseObject::operator delete(void* address) {
-		se::memory::_delete(address);
-	}
-
-	const auto BaseObject_dtor = reinterpret_cast<TES3::BaseObject * (__thiscall*)(TES3::BaseObject*)>(0x4F0CA0);
-	void BaseObject::dtor() {
-		const auto dataHandler = TES3::DataHandler::get();
-
-		clearCachedLuaObject(this);
-		mwse::ReferenceTracker::invalidateObject(this);
-
-		if (objectType == ObjectType::Cell && dataHandler && dataHandler->nonDynamicData) {
-			dataHandler->nonDynamicData->clearCellByNameCache(static_cast<const Cell*>(this));
-		}
-
-		if (UI::MenuInputController::lastTooltipObject == this) {
-			UI::MenuInputController::lastTooltipObject = nullptr;
-			UI::MenuInputController::lastTooltipItemData = nullptr;
-			UI::MenuInputController::lastTooltipCount = 0;
-		}
-
-		using namespace mwse::lua::event;
-		if (ObjectCopiedEvent::ms_LastCopied == this || ObjectCopiedEvent::ms_LastCopiedFrom == this) {
-			ObjectCopiedEvent::ms_LastCopied = nullptr;
-			ObjectCopiedEvent::ms_LastCopiedFrom = nullptr;
-		}
-
-		BaseObject_dtor(this);
-	}
-
-	bool BaseObject::getObjectModified() const {
-		return (objectFlags & TES3::ObjectFlag::Modified);
-	}
-
-	void BaseObject::setObjectModified(bool modified) {
-		vTable.base->setObjectModified(this, modified);
-	}
-
-	const char* BaseObject::getObjectID() const {
-		return vTable.base->getObjectID(this);
-	}
-
-	const auto BaseObject_writeFileHeader = reinterpret_cast<bool(__thiscall*)(const BaseObject*, GameFile*)>(0x4EEE60);
-	bool BaseObject::writeFileHeader(GameFile* file) const {
-		return BaseObject_writeFileHeader(this, file);
-	}
-
-	bool BaseObject::supportsActivate() const {
-		// Make sure we aren't dealing with references.
-		auto asBase = getBaseObject();
-
-		if (asBase->isItem()) {
-			return static_cast<const Item*>(this)->getIsCarriable();
-		}
-
-		if (asBase->objectType == ObjectType::NPC || asBase->objectType == ObjectType::Creature) {
-			const auto macp = WorldController::get() ? WorldController::get()->getMobilePlayer() : nullptr;
-			if (macp) {
-				return macp->getFlagInCombat();
-			}
-			else {
-				return true;
-			}
-		}
-
-		switch (asBase->objectType) {
-		case TES3::ObjectType::Activator:
-		case TES3::ObjectType::Container:
-		case TES3::ObjectType::Door:
-			return true;
-		default:
-			return false;
-		}
-	}
-
-	BaseObject* BaseObject::getBaseObject() {
-		auto object = static_cast<BaseObject*>(this);
-
-		if (object->objectType == ObjectType::Reference) {
-			object = static_cast<Reference*>(object)->baseObject;
-		}
-
-		if (object->isActor() && static_cast<Actor*>(object)->isClone()) {
-			object = static_cast<Actor*>(object)->getBaseActor();
-		}
-
-		return object;
-	}
-
-	BaseObject const* BaseObject::getBaseObject() const {
-		auto object = static_cast<const BaseObject*>(this);
-
-		if (object->objectType == ObjectType::Reference) {
-			object = static_cast<const Reference*>(object)->baseObject;
-		}
-
-		if (object == nullptr) {
-			return nullptr;
-		}
-
-		if (object->isActor() && static_cast<const Actor*>(object)->isClone()) {
-			object = static_cast<const Actor*>(object)->getBaseActor();
-		}
-
-		return object;
-	}
-
-	bool BaseObject::isPhysicalObject() const {
-		switch (objectType) {
-		case TES3::ObjectType::Activator:
-		case TES3::ObjectType::Alchemy:
-		case TES3::ObjectType::Ammo:
-		case TES3::ObjectType::Apparatus:
-		case TES3::ObjectType::Armor:
-		case TES3::ObjectType::Bodypart:
-		case TES3::ObjectType::Book:
-		case TES3::ObjectType::Clothing:
-		case TES3::ObjectType::Container:
-		case TES3::ObjectType::Creature:
-		case TES3::ObjectType::CreatureClone:
-		case TES3::ObjectType::Door:
-		case TES3::ObjectType::Ingredient:
-		case TES3::ObjectType::LeveledCreature:
-		case TES3::ObjectType::LeveledItem:
-		case TES3::ObjectType::Light:
-		case TES3::ObjectType::Lockpick:
-		case TES3::ObjectType::Misc:
-		case TES3::ObjectType::NPC:
-		case TES3::ObjectType::NPCClone:
-		case TES3::ObjectType::Probe:
-		case TES3::ObjectType::Repair:
-		case TES3::ObjectType::Static:
-		case TES3::ObjectType::Weapon:
-			return true;
-		default:
-			return false;
-		}
-	}
-
-	PhysicalObject* BaseObject::asPhysicalObject() {
-		if (!isPhysicalObject()) return nullptr;
-		return static_cast<PhysicalObject*>(this);
-	}
-
-	const PhysicalObject* BaseObject::asPhysicalObject() const {
-		if (!isPhysicalObject()) return nullptr;
-		return static_cast<const PhysicalObject*>(this);
-	}
-
-	bool BaseObject::isActor() const {
-		switch (objectType) {
-		case TES3::ObjectType::Container:
-		case TES3::ObjectType::Creature:
-		case TES3::ObjectType::CreatureClone:
-		case TES3::ObjectType::NPC:
-		case TES3::ObjectType::NPCClone:
-			return true;
-		default:
-			return false;
-		}
-	}
-
-	bool BaseObject::isMobileCapableActor() const {
-		switch (objectType) {
-		case TES3::ObjectType::Creature:
-		case TES3::ObjectType::CreatureClone:
-		case TES3::ObjectType::NPC:
-		case TES3::ObjectType::NPCClone:
-			return true;
-		default:
-			return false;
-		}
-	}
-
-	bool BaseObject::isItem() const {
-		switch (objectType) {
-		case TES3::ObjectType::Alchemy:
-		case TES3::ObjectType::Ammo:
-		case TES3::ObjectType::Apparatus:
-		case TES3::ObjectType::Armor:
-		case TES3::ObjectType::Book:
-		case TES3::ObjectType::Clothing:
-		case TES3::ObjectType::Ingredient:
-		case TES3::ObjectType::Light:
-		case TES3::ObjectType::Lockpick:
-		case TES3::ObjectType::Misc:
-		case TES3::ObjectType::Probe:
-		case TES3::ObjectType::Repair:
-		case TES3::ObjectType::Weapon:
-			return true;
-		default:
-			return false;
-		}
-	}
-
-	bool BaseObject::isWeaponOrAmmo() const {
-		switch (objectType) {
-		case TES3::ObjectType::Ammo:
-		case TES3::ObjectType::Weapon:
-			return true;
-		default:
-			return false;
-		}
-	}
-
-	const char* BaseObject::getSourceFilename() const {
-		if (sourceMod) {
-			return sourceMod->filename;
-		}
-		return nullptr;
-	}
-
-	bool BaseObject::getLinksResolved() const {
-		return BIT_TEST(objectFlags, TES3::ObjectFlag::LinksResolvedBit);
-	}
-
-	void BaseObject::setLinksResolved(bool value) {
-		BIT_SET(objectFlags, TES3::ObjectFlag::LinksResolvedBit, value);
-	}
-
-	bool BaseObject::getDisabled() const {
-		return BIT_TEST(objectFlags, TES3::ObjectFlag::DisabledBit);
-	}
-
-	bool BaseObject::getDeleted() const {
-		return BIT_TEST(objectFlags, TES3::ObjectFlag::DeleteBit);
-	}
-
-	void BaseObject::setDeleted(bool deleted) {
-		BIT_SET(objectFlags, TES3::ObjectFlag::DeleteBit, deleted);
-	}
-
-	bool BaseObject::getPersistent() const {
-		return BIT_TEST(objectFlags, TES3::ObjectFlag::PersistentBit);
-	}
-
-	void BaseObject::setPersistent(bool value) {
-		BIT_SET(objectFlags, TES3::ObjectFlag::PersistentBit, value);
-	}
-
-	bool BaseObject::getBlocked() const {
-		return BIT_TEST(objectFlags, TES3::ObjectFlag::BlockedBit);
-	}
-
-	void BaseObject::setBlocked(bool value) {
-		BIT_SET(objectFlags, TES3::ObjectFlag::BlockedBit, value);
-	}
-
-	bool BaseObject::getUpdatesCollisionGroups() const {
-		const auto baseObject = getBaseObject();
-
-		switch (baseObject->objectType) {
-		case ObjectType::Activator:
-		case ObjectType::Container:
-		case ObjectType::Door:
-		case ObjectType::Land:
-		case ObjectType::Static:
-			return true;
-		case ObjectType::Light:
-			return !static_cast<const Light*>(baseObject)->getCanCarry();
-		}
-		return false;
-	}
-
-	bool BaseObject::getSupportsLuaData() const {
-		// Gold does all kinds of funky things. No ItemData creation on it is allowed.
-		if (objectType == ObjectType::Misc && static_cast<const Misc*>(this)->isGold()) {
-			return false;
-		}
-
-		// Projectiles cannot have custom data, it breaks the equip interface.
-		if (objectType == ObjectType::Weapon || objectType == ObjectType::Ammo) {
-			return !static_cast<const Weapon*>(this)->isProjectile();
-		}
-
-		return true;
-	}
-
-	std::string BaseObject::toJson() const {
-		std::ostringstream ss;
-		ss << "\"tes3baseObject:" << getObjectID() << "\"";
-		return ss.str();
-	}
-
-	std::unordered_set<const BaseObject*> sourcelessObjects;
-
-	bool BaseObject::getSourceless() const {
-		return isSourcelessObject(this);
-	}
-
-	void BaseObject::setSourceless(bool sourceless) const {
-		if (sourceless) {
-			setSourcelessObject(this);
-		}
-		else {
-			sourcelessObjects.erase(this);
-		}
-	}
-
-	const auto TES3_isSourcelessObject = reinterpret_cast<bool(__stdcall*)(const BaseObject*)>(0x4C1980);
-	bool __stdcall BaseObject::isSourcelessObject(const BaseObject* object) {
-		return TES3_isSourcelessObject(object) || sourcelessObjects.contains(object);
-	}
-
-	void BaseObject::setSourcelessObject(const BaseObject* object) {
-		if (!TES3_isSourcelessObject(object)) {
-			sourcelessObjects.insert(object);
-		}
-	}
-
-	static std::unordered_map<const BaseObject*, sol::object> baseObjectCache;
-
-	bool BaseObject::hasCachedLuaObject() const {
-		return baseObjectCache.contains(this);
-	}
-
-	sol::object BaseObject::getCachedLuaObject() const {
-		const auto stateHandle = mwse::lua::LuaManager::getInstance().getThreadSafeStateHandle();
-		auto cacheHit = baseObjectCache.find(this);
-		if (cacheHit != baseObjectCache.end()) {
-			auto result = cacheHit->second;
-			return result;
-		}
-		return sol::nil;
-	}
-
-	sol::object BaseObject::getOrCreateLuaObject(lua_State* L) const {
-		if (this == nullptr) {
-			return sol::nil;
-		}
-
-		const auto stateHandle = mwse::lua::LuaManager::getInstance().getThreadSafeStateHandle();
-
-		auto cacheHit = baseObjectCache.find(this);
-		if (cacheHit != baseObjectCache.end()) {
-			auto result = cacheHit->second;
-			return result;
-		}
-
-		// Make sure we're looking at the main state.
-		L = stateHandle.getState();
-
-		sol::object ref = sol::nil;
-		switch ((uint32_t)vTable.object) {
-		case TES3::VirtualTableAddress::Activator:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Activator*>(this));
-			break;
-		case TES3::VirtualTableAddress::Alchemy:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Alchemy*>(this));
-			break;
-		case TES3::VirtualTableAddress::AnimationGroup:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::AnimationGroup*>(this));
-			break;
-		case TES3::VirtualTableAddress::Apparatus:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Apparatus*>(this));
-			break;
-		case TES3::VirtualTableAddress::Armor:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Armor*>(this));
-			break;
-		case TES3::VirtualTableAddress::Birthsign:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Birthsign*>(this));
-			break;
-		case TES3::VirtualTableAddress::BodyPart:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::BodyPart*>(this));
-			break;
-		case TES3::VirtualTableAddress::Book:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Book*>(this));
-			break;
-		case TES3::VirtualTableAddress::Cell:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Cell*>(this));
-			break;
-		case TES3::VirtualTableAddress::Class:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Class*>(this));
-			break;
-		case TES3::VirtualTableAddress::Clothing:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Clothing*>(this));
-			break;
-		case TES3::VirtualTableAddress::ContainerBase:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Container*>(this));
-			break;
-		case TES3::VirtualTableAddress::ContainerInstance:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::ContainerInstance*>(this));
-			break;
-		case TES3::VirtualTableAddress::CreatureBase:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Creature*>(this));
-			break;
-		case TES3::VirtualTableAddress::CreatureInstance:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::CreatureInstance*>(this));
-			break;
-		case TES3::VirtualTableAddress::Dialogue:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Dialogue*>(this));
-			break;
-		case TES3::VirtualTableAddress::DialogueInfo:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::DialogueInfo*>(this));
-			break;
-		case TES3::VirtualTableAddress::Door:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Door*>(this));
-			break;
-		case TES3::VirtualTableAddress::Enchantment:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Enchantment*>(this));
-			break;
-		case TES3::VirtualTableAddress::Faction:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Faction*>(this));
-			break;
-		case TES3::VirtualTableAddress::GlobalVariable:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::GlobalVariable*>(this));
-			break;
-		case TES3::VirtualTableAddress::GameSetting:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::GameSetting*>(this));
-			break;
-		case TES3::VirtualTableAddress::Ingredient:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Ingredient*>(this));
-			break;
-		case TES3::VirtualTableAddress::Land:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Land*>(this));
-			break;
-		case TES3::VirtualTableAddress::LeveledCreature:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::LeveledCreature*>(this));
-			break;
-		case TES3::VirtualTableAddress::LeveledItem:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::LeveledItem*>(this));
-			break;
-		case TES3::VirtualTableAddress::Light:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Light*>(this));
-			break;
-		case TES3::VirtualTableAddress::Lockpick:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Lockpick*>(this));
-			break;
-		case TES3::VirtualTableAddress::MagicEffect:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::MagicEffect*>(this));
-			break;
-		case TES3::VirtualTableAddress::Miscellaneous:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Misc*>(this));
-			break;
-		case TES3::VirtualTableAddress::NPCBase:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::NPC*>(this));
-			break;
-		case TES3::VirtualTableAddress::NPCInstance:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::NPCInstance*>(this));
-			break;
-		case TES3::VirtualTableAddress::Probe:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Probe*>(this));
-			break;
-		case TES3::VirtualTableAddress::Quest:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Quest*>(this));
-			break;
-		case TES3::VirtualTableAddress::Race:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Race*>(this));
-			break;
-		case TES3::VirtualTableAddress::Reference:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Reference*>(this));
-			break;
-		case TES3::VirtualTableAddress::Region:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Region*>(this));
-			break;
-		case TES3::VirtualTableAddress::RepairTool:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::RepairTool*>(this));
-			break;
-		case TES3::VirtualTableAddress::Script:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Script*>(this));
-			break;
-		case TES3::VirtualTableAddress::Skill:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Skill*>(this));
-			break;
-		case TES3::VirtualTableAddress::Sound:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Sound*>(this));
-			break;
-		case TES3::VirtualTableAddress::SoundGenerator:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::SoundGenerator*>(this));
-			break;
-		case TES3::VirtualTableAddress::Spell:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Spell*>(this));
-			break;
-		case TES3::VirtualTableAddress::MagicSourceInstance:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::MagicSourceInstance*>(this));
-			break;
-		case TES3::VirtualTableAddress::Static:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Static*>(this));
-			break;
-		case TES3::VirtualTableAddress::Weapon:
-			ref = sol::make_object_userdata(L, static_cast<const TES3::Weapon*>(this));
-			break;
-		}
-
-		if (ref != sol::nil) {
-			baseObjectCache[this] = ref;
-		}
-		else {
-			mwse::log::getLog() << "[MWSE] WARNING: An unknown object type was identified with a virtual table address of 0x" << std::hex << (unsigned int)vTable.base << ". Report this to MWSE developers." << std::endl;
-			mwse::lua::logStackTrace();
-		}
-
-		return ref;
-	}
-
-	void BaseObject::clearCachedLuaObject(const BaseObject* object) {
-		const auto stateHandle = mwse::lua::LuaManager::getInstance().getThreadSafeStateHandle();
-
-		if (!baseObjectCache.empty()) {
-			// Clear any events that make use of this object.
-			auto it = baseObjectCache.find(object);
-			if (it != baseObjectCache.end()) {
-				// Clear any events that make use of this object.
-				mwse::lua::event::clearObjectFilter(it->second);
-
-				// Null the userdata's internal pointer before removing our identity cache.
-				mwse::lua::clearUserdataPointer(it->second);
-
-				// Let people know that this object is invalidated.
-				stateHandle.triggerEvent(new mwse::lua::event::ObjectInvalidatedEvent(it->second));
-
-				// Remove it from the cache.
-				baseObjectCache.erase(it);
-			}
-		}
-	}
-
-	void BaseObject::clearCachedLuaObjects() {
-		const auto stateHandle = mwse::lua::LuaManager::getInstance().getThreadSafeStateHandle();
-		for (auto& item : baseObjectCache) {
-			mwse::lua::clearUserdataPointer(item.second);
-		}
-		baseObjectCache.clear();
-	}
-
 	void Object::copy(const Object* from, bool shareAIPackageConfig) {
 		vTable.object->copy(this, from, shareAIPackageConfig);
 	}
