+++ SharedSE/SharedSE.vcxitems.filters
@@ -31,12 +31,15 @@
     <Filter Include="Source Files\Collections">
       <UniqueIdentifier>{dc2213a8-5f9a-4d9f-8933-eebad18bcffd}</UniqueIdentifier>
     </Filter>
-	<Filter Include="Header Files\CrashLogger">
-	  <UniqueIdentifier>{b29f8a80-21a4-4797-846a-1f06bf75b10b}</UniqueIdentifier>
-	</Filter>
-	<Filter Include="Source Files\CrashLogger">
-	  <UniqueIdentifier>{9d42f240-15aa-44d5-bca9-d5d7ea289b13}</UniqueIdentifier>
-	</Filter>
+    <Filter Include="Header Files\CrashLogger">
+      <UniqueIdentifier>{b29f8a80-21a4-4797-846a-1f06bf75b10b}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\CrashLogger">
+      <UniqueIdentifier>{9d42f240-15aa-44d5-bca9-d5d7ea289b13}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\TES3">
+      <UniqueIdentifier>{27f0c770-3e8f-4c0c-a828-8ed40c82a5ff}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="$(MSBuildThisFileDirectory)NIIteratedList.h">
@@ -345,6 +348,18 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)NIPoint2.h">
       <Filter>Header Files\NI</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3BaseObject.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3Defines.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3GameFile.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3VirtualTableDefines.Morrowind.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
@@ -566,28 +581,34 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)MemoryUtil.cpp">
       <Filter>Source Files\Util</Filter>
     </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)TES3BaseObject.cpp">
+      <Filter>Source Files\TES3</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)TES3GameFile.cpp">
+      <Filter>Source Files\TES3</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
-	<ClInclude Include="$(MSBuildThisFileDirectory)CrashLogger.h">
-	  <Filter>Header Files\CrashLogger</Filter>
-	</ClInclude>
-	<ClInclude Include="$(MSBuildThisFileDirectory)CrashLogUtilities.h">
-	  <Filter>Header Files\CrashLogger</Filter>
-	</ClInclude>
-	<ClCompile Include="$(MSBuildThisFileDirectory)CrashLogger.cpp">
-	  <Filter>Source Files\CrashLogger</Filter>
-	</ClCompile>
-	<ClCompile Include="$(MSBuildThisFileDirectory)CrashLogDevice.cpp">
-	  <Filter>Source Files\CrashLogger</Filter>
-	</ClCompile>
-	<ClCompile Include="$(MSBuildThisFileDirectory)CrashLogMemory.cpp">
-	  <Filter>Source Files\CrashLogger</Filter>
-	</ClCompile>
-	<ClCompile Include="$(MSBuildThisFileDirectory)CrashLogRegistryStack.cpp">
-	  <Filter>Source Files\CrashLogger</Filter>
-	</ClCompile>
-	<ClCompile Include="$(MSBuildThisFileDirectory)CrashLogSections.cpp">
-	  <Filter>Source Files\CrashLogger</Filter>
-	</ClCompile>
+    <ClInclude Include="$(MSBuildThisFileDirectory)CrashLogger.h">
+      <Filter>Header Files\CrashLogger</Filter>
+    </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)CrashLogUtilities.h">
+      <Filter>Header Files\CrashLogger</Filter>
+    </ClInclude>
+    <ClCompile Include="$(MSBuildThisFileDirectory)CrashLogger.cpp">
+      <Filter>Source Files\CrashLogger</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)CrashLogDevice.cpp">
+      <Filter>Source Files\CrashLogger</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)CrashLogMemory.cpp">
+      <Filter>Source Files\CrashLogger</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)CrashLogRegistryStack.cpp">
+      <Filter>Source Files\CrashLogger</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)CrashLogSections.cpp">
+      <Filter>Source Files\CrashLogger</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
