+++ SharedSE/TES3Defines.h
@@ -1,5 +1,13 @@
 #pragma once
 
+#if defined(SE_TARGETS_MW) && SE_TARGETS_MW == 1
+
+#include "TES3VirtualTableDefines.Morrowind.h"
+
+#elif defined(SE_TARGETS_CS) && SE_TARGETS_CS == 1
+
+#endif
+
 namespace TES3 {
 	class MagicEffectController;
 	struct ActionAttachment;
@@ -158,98 +166,6 @@ namespace TES3 {
 	struct WeatherThunder;
 	struct WorldController;
 
-	namespace VirtualTableAddress {
-		enum VirtualTableAddress : unsigned int {
-			Activator = 0x747FC4,
-			ActorAnimController = 0x74ADE8,
-			ActorWearsObjects = 0x7478E0,
-			AIPackageActivate = 0x74ADC4,
-			AIPackageBase = 0x74AD58,
-			AIPackageEscort = 0x74AD98,
-			AIPackageFollow = 0x74ADB0,
-			AIPackageTravel = 0x74AD80,
-			AIPackageWander = 0x74AD6C,
-			Alchemy = 0x749684,
-			AnimatedObject = 0x74A580,
-			AnimationGroup = 0x747880,
-			Apparatus = 0x748110,
-			ArchiveFile = 0x7478BC,
-			Armor = 0x748258,
-			Attribute = 0x746464,
-			AttributeFatigue = 0x746600,
-			AttributeSkill = 0x746AD4,
-			BaseObject = 0x74A2D8,
-			Birthsign = 0x747644,
-			BodyPart = 0x7483A4,
-			Book = 0x7484EC,
-			Cell = 0x74A0F8,
-			Class = 0x74939C,
-			Clothing = 0x748634,
-			ContainerBase = 0x74877C,
-			ContainerInstance = 0x7488FC,
-			CreatureBase = 0x747A74,
-			CreatureInstance = 0x747BF4,
-			CutscenePlayer = 0x7467A0,
-			Dialogue = 0x74983C,
-			DialogueInfo = 0x749818,
-			Door = 0x748A7C,
-			Enchantment = 0x749548,
-			Faction = 0x7497CC,
-			Game = 0x74665C,
-			GameBase = 0x74A920,
-			GameSetting = 0x749868,
-			GlobalVariable = 0x7497F0,
-			Ingredient = 0x748BC8,
-			Land = 0x7498D8,
-			LandTexture = 0x7498B4,
-			LeveledCreature = 0x749958,
-			LeveledItem = 0x749AA8,
-			Light = 0x749BF4,
-			Lockpick = 0x748D10,
-			MagicEffect = 0x7493E4,
-			MagicSourceInstance = 0x74AA14,
-			Miscellaneous = 0x748E58,
-			MobileActor = 0x74AB4C,
-			MobileCreature = 0x74AFA4,
-			MobileNPC = 0x74AE6C,
-			MobileObject = 0x74B0C0,
-			MobilePlayer = 0x74B174,
-			MobileProjectile = 0x74B2B4,
-			NPCBase = 0x749DE8,
-			NPCInstance = 0x749F68,
-			Object = 0x74A444,
-			PathGrid = 0x74A8F8,
-			PhysicalObject = 0x74A2FC,
-			PlayerAnimController = 0x74AE30,
-			Probe = 0x748FA0,
-			Quest = 0x7469E0,
-			Race = 0x749378,
-			Reference = 0x74A140,
-			Region = 0x7477BC,
-			RepairTool = 0x7490E8,
-			Script = 0x74A990,
-			Skill = 0x7493C0,
-			Sound = 0x74A9C0,
-			SoundGenerator = 0x74A9EC,
-			Spell = 0x749408,
-			SpellProjectile = 0x74B360,
-			Static = 0x749230,
-			TArray = 0x747A60,
-			TES3Archive = 0x7478B0,
-			Weapon = 0x74A7A4,
-			WeatherAshstorm = 0x746C60,
-			WeatherBase = 0x746BD4,
-			WeatherBlight = 0x746C70,
-			WeatherBlizzard = 0x746C90,
-			WeatherClear = 0x746BE4,
-			WeatherCloudy = 0x746BFC,
-			WeatherFog = 0x746C0C,
-			WeatherOvercast = 0x746C1C,
-			WeatherRain = 0x746C2C,
-			WeatherSnow = 0x746C80,
-			WeatherStorm = 0x746C48,
-		};
-	}
 
 	namespace UI {
 		struct Element;
@@ -260,4 +176,69 @@ namespace TES3 {
 		class LuaData;
 	}
 
+	//
+	// Object types. They are char[4], or can be interpreted as a 32-bit integer.
+	//
+
+	namespace ObjectType {
+		enum ObjectType {
+			Invalid = 0,
+			Activator = 'ITCA',
+			Alchemy = 'HCLA',
+			Ammo = 'OMMA',
+			AnimationGroup = 'GINA',
+			Apparatus = 'APPA',
+			Armor = 'OMRA',
+			Birthsign = 'NGSB',
+			Bodypart = 'YDOB',
+			Book = 'KOOB',
+			Cell = 'LLEC',
+			Class = 'SALC',
+			Clothing = 'TOLC',
+			Container = 'TNOC',
+			Creature = 'AERC',
+			CreatureClone = 'CERC',
+			Dialogue = 'LAID',
+			DialogueInfo = 'OFNI',
+			Door = 'ROOD',
+			Enchantment = 'HCNE',
+			Faction = 'TCAF',
+			GameSetting = 'TSMG',
+			Global = 'BOLG',
+			Ingredient = 'RGNI',
+			Land = 'DNAL',
+			LandTexture = 'XETL',
+			LeveledCreature = 'CVEL',
+			LeveledItem = 'IVEL',
+			Light = 'HGIL',
+			Lockpick = 'KCOL',
+			MagicEffect = 'FEGM',
+			MagicSourceInstance = 'LLPS',
+			Misc = 'CSIM',
+			MobileCreature = 'RCAM',
+			MobileNPC = 'HCAM',
+			MobileObject = 'TCAM',
+			MobilePlayer = 'PCAM',
+			MobileProjectile = 'JRPM',
+			MobileSpellProjectile = 'PSPM',
+			NPC = '_CPN',
+			NPCClone = 'CCPN',
+			PathGrid = 'DRGP',
+			Probe = 'BORP',
+			Quest = 'SEUQ',
+			Race = 'ECAR',
+			Reference = 'RFER',
+			Region = 'NGER',
+			Repair = 'APER',
+			Script = 'TPCS',
+			Skill = 'LIKS',
+			Sound = 'NUOS',
+			SoundGenerator = 'GDNS',
+			Spell = 'LEPS',
+			Static = 'TATS',
+			TES3 = '3SET',
+			Training = 'IART',
+			Weapon = 'PAEW',
+		};
+	}
 }
